feat: fix checkout script

+++ sd-open-webui/entrypoint.sh
@@ -1,13 +1,17 @@
 #!/bin/bash
 set -x
 
-# reclone the repo
-git clone --depth 1 --branch $1 https://github.com/AUTOMATIC1111/stable-diffusion-webui /workdir/temp/stable-diffusion-webui
-cd /workdir/temp/stable-diffusion-webui
+# reclone the repo if not exists
+if [ ! -d "/workdir/temp/stable-diffusion-webui" ]; then
+  mkdir -p /workdir/temp
+  git clone --depth 1 --branch $1 https://github.com/AUTOMATIC1111/stable-diffusion-webui /workdir/temp/stable-diffusion-webui
+  
+  # copy update the original workdir
+  cp -R -u -p /workdir/temp/stable-diffusion-webui/* /workdir
+  rm -rdf /workdir/temp
+fi
 
-# copy update the original workdir
-cp -R -u -p /workdir/temp/stable-diffusion-webui/* /workdir
-rm -rdf /workdir/temp
+# fix the permissions
 chown root:root -R /workdir
 
 # create virtual environment
